분석

+++ src/main/java/bookAlgorithms/dynamicPrograming/FiboTest.java
@@ -8,14 +8,14 @@ public class FiboTest {
 	@Test
 	public void test1() {
 		int number = 7;
-		int[] array = new int[number+1];
+		long[] array = new long[number+1];
 		assertEquals(13, Fibonazzi.fibonazzi(array, number));
 	}
 
 	@Test
 	public void test2() {
 		int number = 131;
-		int[] array = new int[number+1];
+		long[] array = new long[number+1];
 		assertEquals(34, Fibonazzi.fibonazzi(array, number));
 	}
